Version 9 (#1207)
* remove dotnet (#1151)

* remove rails and ruby support (#1150)

* remove rails and ruby support

* last ruby support

* remove zray (#1152)

* Refactor opt in features to use yaml collections (#1160)

* Utilize collections to install optional packages/features


* Change chronograph install to use features


* Change couchdb install to use features


* Change crystal install to use features


* Change docker install to use features


* Change elasticsearch install to use features


* Change golang install to use features


* Change grafana install to use features


* Change influxdb install to use features


* Change mariadb install to use features


* Change minio install to use features


* Change mongodb install to use features


* Change mysql8 install to use features


* Change neo4j install to use features


* Change ohmyzsh install to use features


* Change python install to use features


* Change webdriver install to use features


* Move opt-in features install up to avoid problems


* Update databases to use the collection of opt-in features


* Default VM name to 'homestead'

* move site type scripts to sub folder

* Remove blackfire block now that it has been moved to features

* Clean up folder Samba credentials Ruby code

* Set script permission to be executable (#1164)


* Group opt-in lockfiles (#1166)

* Create homestead-features folder to hold lockfiles


* Update opt-in features to use homestead-features folder for lockfiles


* Implement Apache Solr as optional feature (#1169)

* Implement Apache Solr as optional feature

* Set script permission to be executable (Solr Feature)

* Use apt instead of apt-get

* Implement Apache Cassandra as a feature (#1170)

* Implement Apache Solr as optional feature

* Set script permission to be executable (Solr Feature)

* Use apt instead of apt-get

* Implement database Apache Cassandra as feature

* Some issues with Java > 8 when running Cassandra

* Add folder check to some features which were missing

* Add more feature folder checks

* Fix bad paste in mongo

* Fix vagrant provision error, fix php restart

* Update alpha2 version

* Require v8 base box

* Fix Graphana install & CouchDB install

* Implement Gearman Job Server as optional feature (#1180)

* Implement Apache Solr as optional feature

* Set script permission to be executable (Solr Feature)

* Use apt instead of apt-get

* Implement Gearman Job Server as optional feature

* Use apt-get for installing gearman job server

* Implement RabbitMQ Message Broker as optional Feature (#1182)

* Implement Apache Solr as optional feature

* Set script permission to be executable (Solr Feature)

* Use apt instead of apt-get

* Implement Gearman Job Server as optional feature

* Use apt-get for installing gearman job server

* Implement RabbitMQ Message Broker as optional Feature

* Forgot to install AMQP PHP Extension for handling RabbitMQ

* PHP AMQP Extension relies on PHP BCMath Extension

* Changed default password for RabbitMQ

* Grouped Repository setup for RabbitMQ Installation

* Update golang version to 1.12.6 (#1183)


* Feature parity from v8.6.0

* Consistency for feature installers

* Add aliases for 5.6 and 7.0, Update Vagrantfile to require recent Vagrant

* Tag beta and update issue template

* Allow configuration of the default PHP version used for the CLI (#1194)

* Don't remove symlink to LVM volume when installing MariaDB

* Small tweaks (#1197)

* Deviantintegral fix xhgui (#1206)

* Correct path for xhgui and mongo scripts

* Bringing in @ deviantintegral's xhgui & mongodb PR & tweaking

* Add NodeJS rollback scripts to after.sh

* version bump

+++ scripts/features/golang.sh
@@ -1,8 +1,17 @@
 #!/usr/bin/env bash
 
+if [ -f /home/vagrant/.homestead-features/go ]
+then
+    echo "Golang already installed."
+    exit 0
+fi
+
+touch /home/vagrant/.homestead-features/go
+chown -Rf vagrant:vagrant /home/vagrant/.homestead-features
+
 # Install Golang
 
-golangVersion="1.12.5"
+golangVersion="1.12.6"
 wget https://dl.google.com/go/go${golangVersion}.linux-amd64.tar.gz -O golang.tar.gz
 tar -C /usr/local -xzf golang.tar.gz go
 printf "\nPATH=\"/usr/local/go/bin:\$PATH\"\n" | tee -a /home/vagrant/.profile
